# LAB/LAB_04/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Film Manager API
  description: |-
    REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3001/
    description: "Local Development Server"
...
paths:
...
  /api/films:
    post:
      tags:
        - api/films
      security:
        - cookieAuth: []
      summary: Create a new film
      description: A new film is created by the authenticated user (who becomes the owner).
      operationId: createFilm
      requestBody:
        description: Representation of the film to be created (with no id because it is assigned by the service)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmCreate'
            examples:
              valid_film_creation:
                value:
                  title: "The Great Adventure"
                  owner: 67890
                  private: false
                  watchDate: "2024-11-09"
                  rating: 8
                  favorite: true
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
              examples:
                film_created_successfully:
                  value:
                    id: 12345
                    title: "The Great Adventure"
                    owner: 67890
                    private: false
                    watchDate: "2024-11-09"
                    rating: 8
                    favorite: true
                    self: "/api/films/12345"
                    update: "/api/films/12345"
                    delete: "/api/films/12345"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Generic bad request error"
                  value:
                    code: 400
                    message: "the request is malformed or invalid. Please check the request data."
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
  /api/films/private:
    get:
      tags:
        - api/films/private
      security:
        - cookieAuth: []
      summary: Retrieve the private films of the logged-in user
      description: |
        The private films of the logged-in user are retrieved. A pagination mechanism is used to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Films'
              examples:
                successful_response:
                  value:
                    totalPages: 5
                    currentPage: 1
                    totalItems: 20
                    films:
                      - id: 123
                        title: "The Private Journey"
                        owner: 67890
                        private: true
                        watchDate: "2024-03-10"
                        rating: 9
                        favorite: true
                        self: "/api/films/private/123"
                        update: "/api/films/private/123"
                        delete: "/api/films/private/123"
                      - id: 124
                        title: "Hidden Secrets"
                        owner: 67890
                        private: true
                        watchDate: "2024-04-22"
                        rating: 7
                        favorite: false
                        self: "/api/films/private/124"
                        update: "/api/films/private/124"
                        delete: "/api/films/private/124"
                    next: "/api/films/private?pageNo=2"
                    previous: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
  /api/films/private/{filmId}:
    get:
      tags:
        - api/films/private/{filmId}
      security:
        - cookieAuth: []
      summary: Retrieve a private film
      description: |
        The private film with ID `filmId` is retrieved. This operation can be performed on the film if the user who performs the operation is the film's owner.
      operationId: getSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
              examples:
                successful_response:
                  value:
                    id: 123
                    title: "The Hidden Life"
                    owner: 67890
                    private: true
                    watchDate: "2023-12-25"
                    rating: 8
                    favorite: true
                    self: "/api/films/private/123"
                    update: "/api/films/private/123"
                    delete: "/api/films/private/123"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '403':
          description: Forbidden (user is not the owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Access is forbidden"
                  value:
                    code: 403
                    message: "you do not have permission to access this resource."
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
    put:
      tags:
        - api/films/private/{filmId}
      security:
        - cookieAuth: []
      summary: Update a private film
      description: |
        The private film with ID `filmId` is updated. This operation does not allow changing its visibility. 
        This operation can be performed only by the owner.
      operationId: updateSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmUpdate'
            examples:
              updateFilmTitleAndRating:
                summary: Update the title and rating of the film
                value:
                  title: "The Hidden Life - Revised Edition"
                  owner: 67890
                  rating: 9
                  favorite: true
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Generic bad request error"
                  value:
                    code: 400
                    message: "the request is malformed or invalid. Please check the request data."
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '403':
          description: Forbidden (user is not the owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Access is forbidden"
                  value:
                    code: 403
                    message: "you do not have permission to access this resource."
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '409':
          description: Conflict (attempt to make unallowed change, e.g., visibility)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Conflict with the current state of the resource"
                  value:
                    code: 409
                    message: "a conflict occurred due to an existing resource or data inconsistency. Please ensure the resource identifiers or data are correct."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
    delete:
      tags:
        - api/films/private/{filmId}
      security:
        - cookieAuth: []
      summary: Delete a private film
      description: |
        The private film with ID `filmId` is deleted. This operation can only be performed by the owner.
      operationId: deleteSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '403':
          description: Forbidden (user is not the owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Access is forbidden"
                  value:
                    code: 403
                    message: "you do not have permission to access this resource."
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
  /api/films/public:
    get:
      tags:
        - api/films/public
      summary: Retrieve the public films
      description: |
        The public films (i.e., the films that are visible for all the users of the service) are retrieved. 
        This operation does not require authentication. A pagination mechanism is used to limit the size of messages.
      operationId: getPublicFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Films'
              examples:
                paginatedFilmsPage1:
                  value:
                    $schema: "string"
                    totalPages: 5
                    currentPage: 1
                    totalItems: 20
                    films:
                      - id: 1
                        title: "Film 1"
                        owner: 123
                        private: false
                        watchDate: "2024-01-01"
                        rating: 8
                        favorite: true
                      - id: 2
                        title: "Film 2"
                        owner: 124
                        private: false
                        watchDate: "2024-02-01"
                        rating: 7
                        favorite: false
                      - id: 3
                        title: "Film 3"
                        owner: 125
                        private: false
                        watchDate: "2024-03-01"
                        rating: 9
                        favorite: true
                    next: "/api/films/public?pageNo=2"
                    previous: null
                paginatedFilmsPage2:
                  value:
                    $schema: "string"
                    totalPages: 5
                    currentPage: 2
                    totalItems: 20
                    films:
                      - id: 11
                        title: "Film A"
                        owner: 123
                        private: false
                        watchDate: "2024-01-15"
                        rating: 8
                        favorite: true
                      - id: 12
                        title: "Film B"
                        owner: 124
                        private: false
                        watchDate: "2024-02-10"
                        rating: 7
                        favorite: false
                      - id: 13
                        title: "Film C"
                        owner: 125
                        private: false
                        watchDate: "2024-03-01"
                        rating: 9
                        favorite: true
                      - id: 14
                        title: "Film D"
                        owner: 126
                        private: false
                        watchDate: "2024-03-10"
                        rating: 6
                        favorite: false
                    next: "/api/films/public?pageNo=3"
                    previous: "/api/films/public?pageNo=1"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
  /api/films/public/invited:
    get:
      tags:
        - api/films/public/invited
      security:
        - cookieAuth: []
      summary: Retrieve the public films that the logged-in user has been invited to review
      description: |
        The public films that the logged-in user has been invited to review are retrieved. A pagination mechanism is used to limit the size of messages.
      operationId: getInvitedFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Films'
              examples:
                invitedFilmsPage2:
                  value:
                    $schema: "string"
                    totalPages: 3
                    currentPage: 2
                    totalItems: 5
                    films:
                      - id: 3
                        title: "Film A"
                        owner: 123
                        private: false
                        watchDate: "2024-01-15"
                        rating: 8
                        favorite: true
                      - id: 4
                        title: "Film B"
                        owner: 124
                        private: false
                        watchDate: "2024-02-10"
                        rating: 7
                        favorite: false
                    next: "/api/films/public/invited?pageNo=3"
                    previous: "/api/films/public/invited?pageNo=1"
                invitedFilmsPage1:
                  value:
                    $schema: "string"
                    totalPages: 3
                    currentPage: 1
                    totalItems: 5
                    films:
                      - id: 1
                        title: "Film 1"
                        owner: 123
                        private: false
                        watchDate: "2024-01-01"
                        rating: 8
                        favorite: true
                      - id: 2
                        title: "Film 2"
                        owner: 124
                        private: false
                        watchDate: "2024-02-01"
                        rating: 7
                        favorite: false
                    next: "/api/films/public/invited?pageNo=2"
                    previous: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
...
  /api/films/public/assignments:
    post:
      tags:
        - api/films/public/assignments
      security:
        - cookieAuth: []
      summary: Assign automatically review invitations for the unassigned public films owned by the authenticated user in a balanced manner
      description: |
        The films owned by the authenticated user who performed the request and that are not assigned to any user for review, are automatically assigned to the users of the service in a balanced manner. The assignments are returned. Only authenticated users can make this request.
      operationId: assignReviewBalanced
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    filmId:
                      type: integer
                    reviewerIds:
                      type: array
                      items:
                        properties:
                          id:
                            type: integer
              examples:
                balancedAssignments:
                  summary: Balanced assignment of reviews to users
                  value:
                    - filmId: 55
                      reviewerIds: [101, 102]
                    - filmId: 56
                      reviewerIds: [103, 104, 105]
                    - filmId: 57
                      reviewerIds: [101, 102, 103]
                balancedNoAssignments:
                  summary: No films available to assign
                  value: []
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
...
components:
  schemas:
...
    FilmCreate:
      description: "schema used to create a new film resource."
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/FilmBase"
      properties:
        private:
          type: boolean
          description: "true if only the owner of the film data structure can access it, false otherwise"
          default: true
      required:
        - private
    FilmUpdate:
      description: "schema used to update an existing film resource."
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/FilmBase"
    Film:
      description: "represents a film resource, including film details and links for interacting with the film (update, delete, etc.)."
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/FilmBase"
      properties:
        id:
          type: integer
          description: "unique identifier for the film data structure"
        private:
          type: boolean
          description: "true if only the owner of the film data structure can access it, false otherwise"
          default: true
        # Links according to HATEOAS principle
        self:
          type: string
          format: uri
          description: "Link to the current film resource"
        update:
          type: string
          format: uri
          description: "Link to update the film resource"
        delete:
          type: string
          format: uri
          description: "Link to delete the film resource"
        reviews:
          type: string
          format: uri
          description: "the reviews for the film"
      required:
        - id
        - private
...
    ErrorResponse:
      description: "Represents an error response returned by the API."
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: "path to the JSON Schema used for validating the error response structure"
        code:
          type: integer
          description: "error code"
        message:
          type: string
          description: "a brief description of the error"
      required:
        - code
        - message
  securitySchemes:
    cookieAuth:
      description: "JWT to identify the authenticated user."
      type: apiKey
      in: cookie
      name: connect.sid